[Silver III] Title: 꿀 아르바이트, Time: 748 ms, Memory: 110224 KB -BaekjoonHub

+++ 백준/Silver/12847. 꿀 아르바이트/꿀 아르바이트.java
@@ -1,4 +1,6 @@
 import java.util.*;
+import java.lang.*;
+import java.io.*;
 
 class Main {
     public static void main(String[] args) {
@@ -7,19 +9,15 @@ class Main {
         int m = sc.nextInt();
         long[] array = new long[n];
         long num = 0;
-
         for(int i = 0; i < n; i++){
-            array[i] = sc.nextLong();
+            array[i] = sc.nextInt();
         }
-
         for(int i = 0; i < m; i++){
             num += array[i];
         }
-
         long maxSum = num;
-
         for(int i = m; i < n; i++){
-            num += array[i] - array[i - m];
+            num += array[i] - array[i-m];
             maxSum = Math.max(maxSum, num);
         }
